LayoutTests/http/tests/security/srcdoc-in-sandbox-cannot-access-parent.html

<script>
if (window.testRunner) {
    testRunner.dumpAsText();
}
</script>
<div id="result">PASS</div>
<iframe sandbox="allow-scripts" srcdoc="
    Trying to access parent...
    <script>
    parent.document.getElementById('result').innerHTML = 'FAIL';
    </script>
"></iframe>
