LayoutTests/http/tests/security/srcdoc-in-sandbox-cannot-access-parent-expected.txt

CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "null" from accessing a frame at "http://127.0.0.1:8000".  The frame requesting access is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: line 4: TypeError: 'undefined' is not an object (evaluating 'parent.document.getElementById')
PASS
